package: move package to each typeˆ
