+++ graphene-writer/src/main/kotlin/com/graphene/writer/store/key/handler/LoggingKeyStoreHandler.kt
@@ -1,4 +1,4 @@
-package com.graphene.writer.store.key
+package com.graphene.writer.store.key.handler
 
 import com.graphene.writer.input.GrapheneMetric
 import com.graphene.writer.store.KeyStoreHandler
