--- graphene-writer/src/main/kotlin/com/graphene/writer/store/key/model/ElasticsearchClientDecorator.kt
@@ -1,5 +1,0 @@
-package com.graphene.writer.store.key.model
-
-abstract class ElasticsearchClientDecorator(
-  val elasticsearchClient: ElasticsearchClient
-) : ElasticsearchClient
