+++ graphene-writer/src/main/kotlin/com/graphene/writer/store/key/property/ElasticsearchKeyStoreHandlerProperty.kt
@@ -1,4 +1,4 @@
-package com.graphene.writer.store.key.model
+package com.graphene.writer.store.key.property
 
 import com.graphene.writer.config.IndexBulkConfiguration
